Rename TwoInteger to TwoInteger*s*
